requirements.txt
aiohttp==3.6.2
amqp==2.5.1
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.5
logger==1.4
MarkupSafe==1.1.1
more-itertools==7.2.0
multidict==4.5.2
pyee==6.0.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
slackclient==2.3.1
slackeventsapi==2.1.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
yarl==1.3.0
zipp==0.6.0
